Fix LWP data type

+++ cloudnetpy/instruments/hatpro.py
@@ -129,7 +129,7 @@ class HatproBin:
         for sample in range(self.header['_n_samples'][0]):
             data['time'][sample] = np.fromfile(file, np.int32, 1)
             data['quality_flag'][sample] = np.fromfile(file, np.int8, 1)
-            data['LWP'][sample] = np.fromfile(file, np.int32, 1)
+            data['LWP'][sample] = np.fromfile(file, np.float32, 1)
             data['_instrument_angles'][sample] = np.fromfile(file, angle_dtype, 1)
 
         data = _add_zenith(version, data)
